feat: errorhandling with validationResult

+++ controllers/dish_controller.js
@@ -14,6 +14,10 @@ module.exports.getAll = async (req, res) => {
 
 // get:id /api/dish/:id
 module.exports.getOne = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { id } = req.params;
     const dish = await prisma.dish.findUnique({
         where: {
@@ -26,6 +30,10 @@ module.exports.getOne = async (req, res) => {
 
 // Search /api/dish/search/:name
 module.exports.search = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { name } = req.params;
     const dish = await prisma.dish.findMany({
         where: {
@@ -43,6 +51,10 @@ module.exports.search = async (req, res) => {
 
 // Create /api/dish
 module.exports.create = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     //const { name } = req.body;
     const dish = await prisma.dish.create({
         data: {
@@ -54,6 +66,10 @@ module.exports.create = async (req, res) => {
 
 // delete /api/dish/:id
 module.exports.delete = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { id } = req.params;
     const dish = await prisma.dish.delete({
         where: {
