+++ controllers/meal_controller.js
@@ -4,6 +4,10 @@ const prisma = new PrismaClient();
 
 // getAll /api/meal
 module.exports.getAll = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const meals = await prisma.meal.findMany({
         include: {
             dish: true
@@ -17,6 +21,10 @@ module.exports.getAll = async (req, res) => {
 
 // Search /api/meal/search/:name
 module.exports.search = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { name } = req.params;
     const meal = await prisma.meal.findMany({
         where: {
@@ -36,6 +44,10 @@ module.exports.search = async (req, res) => {
 
 // Create /api/meal, creates a new meal that a user has eaten
 module.exports.create = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { userId, dishId, type } = req.body;
 
     const meal = await prisma.meal.create({
@@ -50,6 +62,10 @@ module.exports.create = async (req, res) => {
 
 // Update /api/meal/:id, updates a meal that a user has eaten
 module.exports.update = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { id } = req.params;
     const { userId, dishId, type } = req.body;
 
@@ -68,6 +84,10 @@ module.exports.update = async (req, res) => {
 
 // delete /api/meal/:id
 module.exports.delete = async (req, res) => {
+    if (!validationResult(req).isEmpty()) {
+        return res.status(400).json({ errors: validationResult(req).array() });
+    }
+
     const { id } = req.params;
     const meal = await prisma.meal.delete({
         where: {
